Gemini clone converted into react project along with authentication using clerk
