+++ clerk-javascript/gemini.css
@@ -1,0 +1,302 @@
+/* Importing Google fonts - Poppins */
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap');
+
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: "Poppins", sans-serif;
+}
+
+:root{
+    /* Dark Mode Colors */
+    --text-color: #E3E3E3;
+    --subheading-color : #828282;
+    --placeholder-color :#A6A6A6;
+    --primary-color : #242424;
+    --secondary-color: #383838;
+    --secondary-hover-color: #444;
+}
+
+.light_mode{
+       /* Light Mode Colors */
+    --text-color: #222;
+    --subheading-color : #A0A0A0;
+    --placeholder-color :#6C6C6C;
+    --primary-color : #FFF;
+    --secondary-color: #E9EEF6;
+    --secondary-hover-color: #DBE1EA;
+}
+
+body{
+    background: var(--primary-color);
+}
+
+.header, .typing-area{
+    color:var(--text-color);
+}
+
+.header ,.chat-list ,.typing-form{
+    max-width: 1000px;
+    margin: 0 auto;
+    
+}
+
+.header{
+    margin: 7vh;
+    padding:1rem;
+    /* display:none; */
+}
+
+body.hide-header .header{
+    display: none;
+    margin-top:0;
+}
+
+.header :where(.title , .subtitle){
+    font-weight: 500;
+    line-height: 3.6rem;
+}
+
+.header .title{  
+    font-size: 2.7rem;
+    width:fit-content;
+    background-clip: text; 
+    background: linear-gradient(to right,  #4285f4,#d96570);
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
+
+.header .subtitle{
+    font-size: 2.4rem;
+    color:var(--subheading-color);
+}
+.suggestion-list{
+    margin-top: 9.5vh;
+    list-style: none;
+    display: flex;
+    gap: 1.25rem;
+    overflow-x: auto;
+    scrollbar-width: none;
+}
+
+.suggestion-list .suggestion{
+    padding: 1.25rem;
+    cursor:pointer;
+    width: 227px;
+    flex-shrink: 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    border-radius: 0.75rem;
+    background: var(--secondary-color);
+}
+
+
+.suggestion-list .suggestion .text{
+    font-weight: 400;
+}
+
+
+.suggestion-list .suggestion .icon{
+    height: 42px;
+    width: 42px;
+    font-size:1.3rem;
+    margin-top: 2.5rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    background: var(--primary-color);
+}
+
+.chat-list{
+    padding:2rem 1rem 12rem;
+    max-height: 100vh;
+    overflow-y: auto;
+    scrollbar-color: #999 transparent;
+
+}
+.chat-list .message.incoming{
+    margin-top: 1.5rem;
+    /* margin-bottom: 1.5rem; */
+}
+
+.chat-list .message .message-content{
+    /* color : var(--text-color); */
+    display: flex;
+    gap: 1.5rem;
+    width:100%;
+    align-items: center;
+}
+
+.chat-list .message .icon{
+    height:35px;
+    width: 35px;
+    display: flex;
+    align-items: center;
+    cursor:pointer;
+    font-size: 1.25rem;
+    margin-left: 3.5rem;
+    justify-content: center;
+    border-radius: 50%;
+    background: var(--second ary-color);
+    color:var(--text-color);
+    visibility: hidden;
+}
+
+.chat-list .message.loading .text{
+    display: none;
+}
+
+.chat-list .message:not(.loading):hover .icon:not(.hide){
+    visibility: visible;
+}
+
+
+.chat-list .loading-indicator{
+    display:none;
+    width : 100%;
+    gap: 0.8rem;
+    flex-direction: column;
+}
+
+.chat-list .message.loading .loading-indicator{
+    display:flex;
+}
+
+.chat-list .message .icon:hover{
+    background: var(--secondary-color);
+}
+
+.chat-list .message .avatar{
+    width:40px;
+    height :40px;
+    object-fit : cover;
+    border-radius: 50%;
+    align-self :flex-start;
+}
+
+.chat-list .message.loading .avatar{
+    animation: rotate 3s linear infinite;
+}
+
+@keyframes rotate {
+    0% {
+        transform: rotate(0deg);
+    }
+    100%{
+        transform: rotate(360deg);
+    }
+}
+
+.chat-list .message .text{
+    color:var(--text-color);
+}
+
+
+.chat-list .loading-indicator .loading-bar{
+    height:11px;
+    width:100%;
+    border-radius: 0.14rem;
+    background: linear-gradient(to right, #4285f4, var(--primary-color), #4285f4);
+    animation: animate 3s linear infinite;
+}
+
+.chat-list .loading-indicator .loading-bar:last-child{
+    width: 70%;
+}
+
+@keyframes animate {
+    0%{
+        background-position:  -800px 0;
+    }
+    100%{
+        background-position:  800px 0;
+    }
+}
+
+.typing-area{
+    /* margin-left:15px;
+    margin-right: 15px; */
+    position: fixed;
+    width: 100%;
+    bottom:0;
+    padding:1rem;
+    background: var(--primary-color);
+}
+.typing-area :where(.typing-form , .action-buttons){
+    display: flex;
+    gap: 0.75rem;
+}
+
+.typing-area .input-wrapper{
+    height:56px;
+    width:100%;
+    display:flex;
+    position: relative;
+}
+
+.typing-area .typing-input{
+    width: 100%;
+    height: 100%;
+    border:none;
+    outline:none;
+    font-size:1rem;
+    color: var(--text-color);
+    border-radius: 6.2rem;
+    padding: 1.1rem 4rem 1.1rem 1.5rem; 
+    background: var(--secondary-color);
+}
+
+.typing-area .typing-input:focus{
+    background: var(--secondary-hover-color);
+}
+
+.typing-area .typing-input::placeholder{
+    color:var(--placeholder-color);
+
+}
+
+.typing-area .icon{
+    height: 56px;
+    width: 56px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    background: var(--secondary-color);
+}
+
+.typing-area .icon:hover{
+    background: var(--secondary-hover-color)!important;
+}
+
+.typing-area .input-wrapper .icon{
+    position: absolute;
+    right:0;
+    outline:none;
+    border:none;
+    background: none;
+    color:var(--text-color);
+    transform: scale(0);
+    transition: transform 0.2s ease;
+}
+
+.typing-area .input-wrapper .typing-input:valid ~ .icon{
+    transform: scale(1);
+}
+
+.typing-area .disclaimer-text{
+    font-size: 0.85rem;
+    margin-top: 1rem;
+    text-align: center;
+    color: var(--placeholder-color);
+}
+
+#btnsignin{
+    position: absolute;
+    top: 15px;
+    right: 15px;
+}
