+++ clerk-javascript/index.html
@@ -1,0 +1,92 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title> 
+    <!-- Linking google fonts for icon -->
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Rounded:opsz,wght,FILL,GRAD@24,400,0,0" />
+    <link rel="stylesheet"  href="gemini.css">
+</head>
+<body>
+    <div id="app">
+    
+    <div>
+      
+    <header class="header">
+        <!-- Header greeting s -->
+        <h2 class="title">Hello, There</h2>
+        <h4 class="subtitle">How can i Help you today?</h4>
+        <button id="btnsignin"></button>
+
+        <!-- Suggestion List -->
+        <ul class = "suggestion-list">
+            <!-- suggestion 1 -->
+            <li class="suggestion">
+                <h4 class="text">Help me plan a game night with my 5 best friends for under $100</h4>
+                <span class="icon material-symbols-rounded">
+                    draw
+                    </span>
+            </li>
+            <!-- suggestion 2 -->
+            <li class="suggestion">
+                <h4 class="text">Briefly summarize this concept: urban planning</h4>
+                <span class="icon material-symbols-rounded">
+                    draw
+                    </span>
+            </li>
+            <!-- suggestion 3 -->
+            <li class="suggestion">
+                <h4 class="text">Draft an email with a packing list for an upcoming trip</h4>
+                <span class="icon material-symbols-rounded">
+                    draw
+                    </span>
+            </li>
+            <!-- suggestion 4 -->
+            <li class="suggestion">
+                <h4 class="text">Create an image & Bedtime Story For my children</h4>
+                <span class="icon material-symbols-rounded">
+                    image
+                    </span>
+            </li>
+
+
+        </ul>
+        
+    </header>
+    <!-- Chat List -->
+    <div class="chat-list">
+    </div>
+
+    <!-- search icon -->
+    <div class="typing-area"> 
+        <form action="" class="typing-form">
+            <div class="input-wrapper">
+                <input type="text" placeholder="Enter a prompt here" class="typing-input" required>
+                <button class="icon material-symbols-rounded">
+                    send
+                </button>
+            </div>
+            <div class="action-buttons">
+                <span id="toggle-theme-button"  class="icon material-symbols-rounded">
+                    light_mode
+                </span>
+                <span id="delete-chat-button" class="icon material-symbols-rounded">
+                    delete
+                </span>
+            </div>
+        </form>
+        <p class="disclaimer-text">
+            Gemini may display inaccurate info, including about people, so double-check its responses. Your privacy and Gemini Apps
+        </p>
+    </div>
+  </div>
+  </div>
+
+    <script src="gemini.js">
+        
+    </script>
+    <script type="module" src="/main.js"></script>
+    
+</body>
+</html>
