+++ clerk-javascript/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "clerk-javascript",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^5.4.8"
+  },
+  "dependencies": {
+    "@clerk/clerk-js": "^5.26.4"
+  }
+}
