--- gemini.css
@@ -1,180 +1,0 @@
-/* Importing Google fonts - Poppins */
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap');
-
-
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: "Poppins", sans-serif;
-}
-
-:root{
-    /* Dark Mode Colors */
-    --text-color: #E3E3E3;
-    --subheading-color : #828282;
-    --placeholder-color :#A6A6A6;
-    --primary-color : #242424;
-    --secondary-color: #383838;
-    --secondary-hover-color: #444;
-}
-
-body{
-    background: var(--primary-color);
-}
-
-.header, .typing-area{
-    color:var(--text-color);
-}
-
-.header , .typing-form{
-    max-width: 1000px;
-    margin: 0 auto;
-    
-}
-
-.header{
-    margin: 7vh;
-    padding:1rem;
-    /* display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center; */
-}
-
-.header :where(.title , .subtitle){
-    font-weight: 500;
-    line-height: 3.6rem;
-}
-
-.header .title{  
-    font-size: 2.7rem;
-    width:fit-content;
-    background-clip: text; 
-    background: linear-gradient(to right,  #4285f4,#d96570);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-}
-
-.header .subtitle{
-    font-size: 2.4rem;
-    color:var(--subheading-color);
-}
-.suggestion-list{
-    margin-top: 9.5vh;
-    list-style: none;
-    display: flex;
-    gap: 1.25rem;
-    overflow-x: auto;
-    scrollbar-width: none;
-}
-
-.suggestion-list .suggestion{
-    padding: 1.25rem;
-    cursor:pointer;
-    width: 227px;
-    flex-shrink: 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    border-radius: 0.75rem;
-    background: var(--secondary-color);
-}
-
-
-.suggestion-list .suggestion .text{
-    font-weight: 400;
-}
-
-
-.suggestion-list .suggestion .icon{
-    height: 42px;
-    width: 42px;
-    font-size:1.3rem;
-    margin-top: 2.5rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    background: var(--primary-color);
-}
-
-.typing-area{
-    /* margin-left:15px;
-    margin-right: 15px; */
-    position: fixed;
-    width: 100%;
-    bottom:0;
-    padding:1rem;
-    background: var(--primary-color);
-}
-.typing-area :where(.typing-form , .action-buttons){
-    display: flex;
-    gap: 0.75rem;
-}
-
-.typing-area .input-wrapper{
-    height:56px;
-    width:100%;
-    display:flex;
-    position: relative;
-}
-
-.typing-area .typing-input{
-    width: 100%;
-    height: 100%;
-    border:none;
-    outline:none;
-    font-size:1rem;
-    color: var(--text-color);
-    border-radius: 6.2rem;
-    padding: 1.1rem 4rem 1.1rem 1.5rem; 
-    background: var(--secondary-color);
-}
-
-.typing-area .typing-input:focus{
-    background: var(--secondary-hover-color);
-}
-
-
-.typing-area .typing-input::placeholder{
-    color:var(--placeholder-color);
-
-}
-
-.typing-area .icon{
-    height: 56px;
-    width: 56px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    background: var(--secondary-color);
-}
-
-.typing-area .icon:hover{
-    background: var(--secondary-hover-color)!important;
-}
-
-.typing-area .input-wrapper .icon{
-    position: absolute;
-    right:0;
-    outline:none;
-    border:none;
-    background: none;
-    color:var(--text-color);
-    transform: scale(0);
-    transition: transform 0.2s ease;
-}
-
-.typing-area .input-wrapper .typing-input:valid ~ .icon{
-    transform: scale(1);
-}
-
-.typing-area .disclaimer-text{
-    font-size: 0.85rem;
-    margin-top: 1rem;
-    text-align: center;
-    color: var(--placeholder-color);
-}
